Move adaptaive sampler code to another file
